+++ CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/nickderobertis/py-finstmt/compare/v1.2.0...v1.2.1) (2023-02-21)
+
+
+### Bug Fixes
+
+* Handle calculated items in determining series names. Fixes fcf usage ([#97](https://github.com/nickderobertis/py-finstmt/issues/97)) ([e069fba](https://github.com/nickderobertis/py-finstmt/commit/e069fba4066c2588b76a727b39fe0d73e068eb09))
+
+## [1.2.0](https://github.com/nickderobertis/py-finstmt/compare/v1.1.0...v1.2.0) (2023-02-21)
+
+
+### Features
+
+* Set Name of Series from Statement getattr ([#87](https://github.com/nickderobertis/py-finstmt/issues/87)) ([a22bb49](https://github.com/nickderobertis/py-finstmt/commit/a22bb49010c8306f461e5653cd6abda24b14ff8f))
+
+## [1.1.0](https://github.com/nickderobertis/py-finstmt/compare/v1.0.1...v1.1.0) (2023-02-21)
+
+
+### Features
+
+* Round statements, better str output. Fix operations with forecasted statements. Completely rework snapshot testing ([#94](https://github.com/nickderobertis/py-finstmt/issues/94)) ([e5fe87c](https://github.com/nickderobertis/py-finstmt/commit/e5fe87ce2fdca0951ea4ee8311e5041d7d577b3e))
+
 ## [1.0.1](https://github.com/nickderobertis/py-finstmt/compare/v1.0.0...v1.0.1) (2023-01-29)